Upgrade to unified VLCKit 4.0 (from TVVLCKit/MobileVLCKit 3.7)
Swaps the two legacy per-platform pods for the single VLCKit 4.0.0a21
xcframework (VLC 4.0 core) on both iOS and tvOS. Motivation: modern
VideoToolbox HEVC/AV1 decoding, active maintenance, latest-OS support,
and a shot at real HDR output on the Apple TV 4K (unconfirmed — needs
on-device verification; VLC 3.7 tone-mapped HDR to SDR).

PopcornTorrent only depends on GCDWebServer, so nothing pinned us to the
old VLC — the swap is clean.

Migration (VLC 4.0 rewrote the track/event model):
- Podfile + Podfile.lock: VLCKit 4.0.0a21 replaces TVVLCKit/MobileVLCKit.
- Bridging header + imports: <VLCKit/VLCKit.h> / import VLCKit.
- Track selection moved from index-based (currentVideoSubTitleIndex,
  audioTrackNames/Indexes, currentAudioTrackIndex) to the object-based
  tracks API (textTracks/audioTracks + selectedExclusively /
  deselectAllTextTracks). tvOS + iOS pickers keep their (position, name)
  contract, adapted over the new arrays.
- State enum: .ended/.buffering/.esAdded removed — end-of-file is now
  .stopped; buffering UI is driven by the torrent streamer.
- position is now Double (cast to Float for the progress bars);
  jumpForward/Backward take Double seconds; VLCMedia(url:) is failable;
  willPlay/resetEqualizer/equalizerEnabled replaced (equalizer via
  VLCAudioEqualizer preset objects).
- Private setTextRendererFont* selectors called optionally (?) so a
  missing selector no-ops instead of trapping.

Builds clean on tvOS and iOS. NOT yet merged to master — pending
on-device testing (HDR, Atmos, subtitle display/styling, audio-track
switching, buffering, stability on the alpha).

+++ PopcornTime/Extensions/PCTPlayerViewController+MediaPlayer.swift
@@ -38,12 +38,12 @@ extension PCTPlayerViewController {
         }
         
         center.skipForwardCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
-            self.mediaplayer.jumpForward(Int32((event as! MPSkipIntervalCommandEvent).interval))
+            self.mediaplayer.jumpForward((event as! MPSkipIntervalCommandEvent).interval) // VLCKit 4: seconds as Double
             return .success
         }
-        
+
         center.skipBackwardCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
-            self.mediaplayer.jumpBackward(Int32((event as! MPSkipIntervalCommandEvent).interval))
+            self.mediaplayer.jumpBackward((event as! MPSkipIntervalCommandEvent).interval)
             return .success
         }
     }
